wiki-testing/pyproject.toml
[project]
name = "wiki-testing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "requests>=2.31.0",
    "textblob>=0.19.0",
    "wikipedia-api>=0.8.1",
]
